Move prime.css and theme.js so playwright doesn't complain
